Rename 'files' to 'file'
