+++ backend/yama/export/models.py
@@ -5,7 +5,7 @@ from typing import Any, Self, assert_never
 
 from pydantic import Field, TypeAdapter, model_validator
 
-from yama.files._models import FilePath, FilePathAdapter
+from yama.file._models import FilePath, FilePathAdapter
 from yama.model.models import ModelBase
